Добавлены модели CardAccess ThemeAccess позволяющие выдавать к приватной записи/теме доступ + работа с доступом
Если пользователь пытаеться открыть приватную тему/карточку проверяеться, являеться ли он автором или имеет ли доступ.

Не авторы не могу удалять и добовлять карточки темы.

TODO(В будущем можно будет продумать, уровни доступа, пока что все категарично: нет доступа, нет прав.)

+++ mindkeeper/main/views.py
@@ -14,19 +14,17 @@ class IndexTemplateView(TemplateView):
             context['themes'] = Themes.get_super_themes_by_user(self.request.user)
         return context
 
-# TODO(сделать дать возможность делать темы и карты приватными)
-# TODO(сделать что бы показывались только карточки пользователя в профиле)
-
 
 @login_required
 def super_theme_and_card_list_view(request):
     if request.user.is_authenticated:
         return render(request, "main/catalog.html", {
-                "themes": Themes.get_super_themes_by_user(request.user),
-                "cards": Cards.get_super_cards_by_user(request.user)
-            })
+            "themes": Themes.get_super_themes_by_user(request.user),
+            "cards": Cards.get_super_cards_by_user(request.user)
+        })
     else:
         return render(request, "main/catalog.html")
+
 
 # class SuperThemeListView(ListView):
 #     model = Themes
@@ -39,39 +37,71 @@ def super_theme_and_card_list_view(request):
 
 
 @login_required
-def open_user_theme(request, theme):
-    # TODO(if is_private)
+def open_theme(request, theme):
+    father_theme = theme
+    theme = Themes.objects.filter(pk=theme).first()
+
     context = {
-        'theme': Themes.objects.filter(pk=theme).first,
-        "themes": Themes.objects.filter(sub_theme_to=theme).filter(user=request.user),
+        'father_theme': Themes.objects.filter(pk=father_theme).first,
+        "themes": Themes.objects.filter(sub_theme_to=theme),
         "cards": Cards.objects.filter(theme=theme)
     }
+
+    if theme.is_private:
+        if request.user != theme.user:
+
+            if request.user not in theme.users_with_access():
+                context = {
+                    'message': 'у вас нет доступа, запросить: # TODO()'  # TODO()
+                }
 
     return render(request, "main/catalog.html", context)
 
 
 @login_required
-def open_user_card(request, card,):
-    # TODO(if is_private)
-    context = {
-        "card": Cards.objects.filter(pk=card).first
-    }
+def open_card(request, card):
+    card = Cards.objects.filter(pk=card).first()
+    context = {'card': card}
+
+    if card.is_private:
+        if request.user != card.user:
+            if request.user not in card.users_with_access():
+                context = {
+                    'message': 'у вас нет доступа, запросить: # TODO()'  # TODO()
+                }
 
     return render(request, "main/card.html", context)
 
 
 @login_required
 def add_card_form(request, theme=None):
-    form = AddCardForm()
-    return render(request, "main/add_card.html", {'form': form, 'theme': theme,
-                                                    'action': 'Добавить карточку',
-                                                    })
+    if request.user != Cards.objects.get(theme=theme).user:
+        print('доступа нет')
+        if request.META.get('HTTP_REFERER', False):
+            return HttpResponseRedirect(request.META['HTTP_REFERER'])
+        else:
+            return redirect('main:index')
+
+    else:
+        form = AddCardForm()
+        return render(request, "main/add_card.html", {'form': form, 'theme': theme,
+                                                      'action': 'Добавить карточку',
+                                                      'HTTP_REFERER': request.META['HTTP_REFERER']})
 
 
 @login_required
 def add_theme_form(request, theme=None):
-    form = AddThemeForm()
-    return render(request, "main/add_theme.html", {'form': form, 'theme': theme, 'action': 'Добавить тему'})
+    if request.user != Themes.objects.get(pk=theme).user:
+        print('доступа нет')
+        if request.META.get('HTTP_REFERER', False):
+            return HttpResponseRedirect(request.META['HTTP_REFERER'])
+        else:
+            return redirect('main:index')
+
+    else:
+        form = AddThemeForm()
+        return render(request, "main/add_theme.html", {'form': form, 'theme': theme, 'action': 'Добавить тему',
+                                                       'HTTP_REFERER': request.META['HTTP_REFERER']})
 
 
 @login_required
@@ -87,8 +117,7 @@ def add_card(request):
         title=request.POST['title'], content=request.POST['content']
     )
 
-    # TODO(редиректить обратно)
-    return redirect('main:index')
+    return HttpResponseRedirect(request.POST['HTTP_REFERER'])
 
 
 @login_required
@@ -103,14 +132,19 @@ def add_theme(request):
         title=request.POST['title'],
         sub_theme_to=theme
     )
-    return redirect('main:index')
+
+    return HttpResponseRedirect(request.POST['HTTP_REFERER'])
 
 
 @login_required
 def del_theme(request, theme):
     theme = Themes.objects.filter(pk=theme).first()
     if theme:
-        theme.delete()
+        if theme.user == request.user:
+            theme.delete()
+
+        else:
+            print("Нет доступа")
 
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
@@ -119,11 +153,16 @@ def del_theme(request, theme):
 def del_card(request, card):
     card = Cards.objects.filter(pk=card).first()
     if card:
-        card.delete()
+        if card.user == request.user:
+            card.delete()
+
+        else:
+            print("Нет доступа")
+
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
 
 def global_search(request):
-    # TODO()
+    # TODO(Глобальный динамический поиск)
     print(request.GET)
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
